requirements.txt
# Core dependencies
rawpy>=0.17.1
imageio>=2.28.1
Pillow>=10.0.0
tqdm>=4.65.0

# Development dependencies (optional)
# mypy>=1.3.0
# flake8>=6.0.0
# black>=23.3.0
